Move backgrounds endpoints into separate file
