+++ src/endpoints/backgrounds.js
@@ -1,0 +1,92 @@
+const fs = require('fs');
+const path = require('path');
+const express = require('express');
+const sanitize = require('sanitize-filename');
+const writeFileAtomicSync = require('write-file-atomic').sync;
+
+const { jsonParser, urlencodedParser } = require('../express-common');
+const { DIRECTORIES, UPLOADS_PATH } = require('../constants');
+const { invalidateThumbnail } = require('./thumbnails');
+const { getImages } = require('../util');
+
+let response_dw_bg;
+
+const router = new express.Router();
+
+router.post('/all', jsonParser, function (request, response) {
+    var images = getImages('public/backgrounds');
+    response.send(JSON.stringify(images));
+
+});
+
+router.post('/set', jsonParser, function (request, response) {
+    try {
+        const bg = `#bg1 {background-image: url('../backgrounds/${request.body.bg}');}`;
+        writeFileAtomicSync('public/css/bg_load.css', bg, 'utf8');
+        response.send({ result: 'ok' });
+    } catch (err) {
+        console.log(err);
+        response.send(err);
+    }
+});
+
+router.post('/delete', jsonParser, function (request, response) {
+    if (!request.body) return response.sendStatus(400);
+
+    if (request.body.bg !== sanitize(request.body.bg)) {
+        console.error('Malicious bg name prevented');
+        return response.sendStatus(403);
+    }
+
+    const fileName = path.join('public/backgrounds/', sanitize(request.body.bg));
+
+    if (!fs.existsSync(fileName)) {
+        console.log('BG file not found');
+        return response.sendStatus(400);
+    }
+
+    fs.rmSync(fileName);
+    invalidateThumbnail('bg', request.body.bg);
+    return response.send('ok');
+});
+
+router.post('/rename', jsonParser, function (request, response) {
+    if (!request.body) return response.sendStatus(400);
+
+    const oldFileName = path.join(DIRECTORIES.backgrounds, sanitize(request.body.old_bg));
+    const newFileName = path.join(DIRECTORIES.backgrounds, sanitize(request.body.new_bg));
+
+    if (!fs.existsSync(oldFileName)) {
+        console.log('BG file not found');
+        return response.sendStatus(400);
+    }
+
+    if (fs.existsSync(newFileName)) {
+        console.log('New BG file already exists');
+        return response.sendStatus(400);
+    }
+
+    fs.renameSync(oldFileName, newFileName);
+    invalidateThumbnail('bg', request.body.old_bg);
+    return response.send('ok');
+});
+
+router.post('/upload', urlencodedParser, function (request, response) {
+    response_dw_bg = response;
+    if (!request.body || !request.file) return response.sendStatus(400);
+
+    const img_path = path.join(UPLOADS_PATH, request.file.filename);
+    const filename = request.file.originalname;
+
+    try {
+        fs.copyFileSync(img_path, path.join('public/backgrounds/', filename));
+        invalidateThumbnail('bg', filename);
+        response_dw_bg.send(filename);
+        fs.unlinkSync(img_path);
+    } catch (err) {
+        console.error(err);
+        response_dw_bg.sendStatus(500);
+    }
+});
+
+module.exports = { router };
